Create FS(A1-Batch)D21P2.java
Mr. Alan, the professor in computer science, gives an assignment to his student 
Mark. He gives a list of values and ask Mark to sort the list based on the 
number of one’s in the binary representation of each value. If there are more 
than one value which has same number of one’s, the sorting for those values, 
should be in the original order of list. 
Your task is to help Mark in his task. 

Note: Each value in the list will be in the range of 1 to 10000.

Sample test case

input=5
1 2 3 4 5
output=[1, 2, 4, 3, 5]